Rename NLL scorers files
